# [Nim Con] speedup with default hash