--- a/nim_con/buildopt.sh
+++ b/nim_con/buildopt.sh
@@ -2,6 +2,11 @@
 
 set -eo pipefail
 
+echo
+echo "origin: $(git remote get-url origin)"
+echo "branch: $(git rev-parse --abbrev-ref HEAD)"
+echo "commit: $(git rev-parse HEAD | cut -c1-7)"
+
 if [[ ${DANGER} = true ]]; then
     build_kind=danger
 else
@@ -20,19 +25,20 @@ fi
 # compiler=clang
 
 echo
-echo "\${compiler} = ${compiler}"
+echo "\$compiler=${compiler}"
 echo "${compiler}" --version
 "${compiler}" --version
-echo
 
 rm -rf default*.prof* nimcache
 
+echo
 echo "Compiling profiled executable"
 nim c -d:${build_kind} \
       --cc:"${compiler}" \
       -d:profileGen \
       related_con.nim
 
+echo
 echo "Generating profile"
 cd ..
 cp posts.json posts_orig.json
@@ -50,6 +56,7 @@ if [[ "${compiler}" = *"clang"* ]]; then
     fi
 fi
 
+echo
 echo "Compiling optimized executable"
 nim c -d:${build_kind} \
       --cc:"${compiler}" \