--- a/nim_con/related_con.nimble
+++ b/nim_con/related_con.nimble
@@ -5,5 +5,4 @@ license       = "MIT"
 
 requires "nim >= 2.0.0",
          "jsony#head",
-         "taskpools#head",
-         "xxhash"
+         "taskpools#head"