I found there's a modest but consistent performance improvement if I drop `xxhash` and (automatically) use Nim's default hash for `TagMap`, *and* switch back to clang.

Now, `xxhash` is definitely faster than Nim's default hash in a straight benchmark. But for this code, the combination of inlining hints, the default hash, *and* clang gives a better result.