@michaelsbradleyjr 2023-10-24
See #363.

If I init `Taskpool` before the timer starts, there is another small but consistent performance increase, or at least it seems the results are less bouncy.

I didn't want to make that change in this PR, though, because the perf gains here are independent of making that change.