@jinyus 2023-10-25
Main:
`Nim Concurrent | 18.66 ms | 242.38 ms | 2.14 s`

This PR:
```
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 16.976ms
        Processing time (w/o IO): 18.087ms
        Processing time (w/o IO): 16.741ms
        Processing time (w/o IO): 15.705ms
        Processing time (w/o IO): 15.64ms
        Processing time (w/o IO): 15.682ms
        Processing time (w/o IO): 16.759ms
        Processing time (w/o IO): 17.747ms
        Processing time (w/o IO): 15.688ms
        Processing time (w/o IO): 17.823ms
        Processing time (w/o IO): 15.652ms
        Processing time (w/o IO): 17.705ms
        Processing time (w/o IO): 15.616ms
          Time (mean ± σ):      43.2 ms ±   0.9 ms    [User: 79.9 ms, System: 6.9 ms]
          Range (min … max):    41.8 ms …  44.3 ms    10 runs
         
Checking output
Success: related_posts_nim_con.json is valid!
Generate 20000 posts
Run Benchmark (15k posts)
Running Nim Concurrent
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 222.428ms
        Processing time (w/o IO): 222.992ms
        Processing time (w/o IO): 223.972ms
          Time (mean ± σ):     329.6 ms ±   0.9 ms    [User: 950.2 ms, System: 29.5 ms]
          Range (min … max):   328.9 ms … 330.3 ms    2 runs
         
Generate 60000 posts
Run Benchmark (30k posts)
Running Nim Concurrent
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 1988.582ms
        Processing time (w/o IO): 1990.468ms
        Processing time (w/o IO): 1989.623ms
          Time (mean ± σ):      2.349 s ±  0.009 s    [User: 8.173 s, System: 0.089 s]
          Range (min … max):    2.343 s …  2.355 s    2 runs
```