@jinyus 2023-10-25
> See also #363.
> 
> If I init `Taskpool` before the timer starts, there is another small but consistent performance boost, or at least the results seem less bouncy.
> 
> I did _**not**_ change it in this PR, though, because the perf gains here are independent of making that change.

Only json decoding should be done before timer starts.